backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
fastapi==0.116.2
ffmpeg-python==0.2.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
